Documentation and better naming
