+++ sdk/servicebus/Azure.Messaging.ServiceBus/src/Amqp/AmqpMessageExtensions.cs
@@ -104,7 +104,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
         {
             if (message.Body.TryGetData(out IEnumerable<ReadOnlyMemory<byte>> dataBody))
             {
-                return BinaryData.FromBytes(BodyMemory.FromReadOnlyMemory(dataBody));
+                return BinaryData.FromBytes(Body.FromReadOnlyMemorySegments(dataBody));
             }
             throw new NotSupportedException($"{message.Body.BodyType} cannot be retrieved using the {nameof(ServiceBusMessage.Body)} property." +
                 $"Use {nameof(ServiceBusMessage.GetRawAmqpMessage)} to access the underlying Amqp Message object.");
