+++ sdk/servicebus/Azure.Messaging.ServiceBus/src/Primitives/ServiceBusMessage.cs
@@ -44,7 +44,7 @@ namespace Azure.Messaging.ServiceBus
         /// <param name="body">The payload of the message in bytes.</param>
         public ServiceBusMessage(ReadOnlyMemory<byte> body)
         {
-            AmqpMessageBody amqpBody = new AmqpMessageBody(BodyMemory.FromReadOnlyMemory(body));
+            AmqpMessageBody amqpBody = new AmqpMessageBody(Amqp.Body.FromReadOnlyMemorySegment(body));
             AmqpMessage = new AmqpAnnotatedMessage(amqpBody);
         }
 
@@ -68,7 +68,7 @@ namespace Azure.Messaging.ServiceBus
                 throw new NotSupportedException($"{receivedMessage.AmqpMessage.Body.BodyType} is not a supported message body type.");
             }
 
-            AmqpMessageBody body = new AmqpMessageBody(BodyMemory.FromReadOnlyMemory(dataBody));
+            AmqpMessageBody body = new AmqpMessageBody(Amqp.Body.FromReadOnlyMemorySegments(dataBody));
             AmqpMessage = new AmqpAnnotatedMessage(body);
 
             // copy properties
@@ -140,7 +140,7 @@ namespace Azure.Messaging.ServiceBus
             get => AmqpMessage.GetBody();
             set
             {
-                AmqpMessage.Body = new AmqpMessageBody(BodyMemory.FromReadOnlyMemory(value));
+                AmqpMessage.Body = new AmqpMessageBody(Amqp.Body.FromReadOnlyMemorySegment(value));
             }
         }
